move the web search tool (which uses tavily mcp) to utils/tools module.
